Move the preprocessing step into its own package
